Move some more files around for clarity.
